Fixed IIAFE
