+++ scraper/package.json
@@ -12,6 +12,7 @@
     "jquery": "^3.5.0",
     "jsonfile": "^6.0.1",
     "lodash": "^4.17.15",
+    "loglevel": "^1.7.1",
     "moment": "^2.27.0",
     "natural": "^2.1.5",
     "node-sass": "^4.14.1",
